requirements.txt
# requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
sse-starlette==2.1.0
fastapi
firebase-admin
pydantic
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
groq
PyPDF2
pdfplumber
python-docx
redis==5.0.1
requests==2.31.0


sounddevice 
numpy 
webrtcvad 
faster-whisper 
TTS
# edge-tts
cerebras_cloud_sdk 
huggingface_hub[hf_xet]
twilio